Version 1.8.1

+++ org/team217/rev/CANSparkMax.java
@@ -7,8 +7,8 @@ package org.team217.rev;
  */
 public class CANSparkMax extends com.revrobotics.CANSparkMax {
 
-    protected double resetPosition = 0.0;
-    protected double invertEnc = 1.0;
+    protected double resetPosition = 0;
+    protected int invertEnc = 1;
     
     /**
      * Constructor for creating a {@code CANSparkMax} motor controller for FRC.
@@ -42,7 +42,7 @@ public class CANSparkMax extends com.revrobotics.CANSparkMax {
 	 *        {@code true} if the encoder value should be multiplied by -1
 	 */
 	public void invertEncoder(boolean isInverted) {
-		invertEnc = (isInverted) ? -1.0 : 1.0;
+		invertEnc = (isInverted) ? -1 : 1;
     }
     
     /** Gets the encoder velocity, modified so {@code invertEncoder()}, {@code setEncoder()}, and {@code resetEncoder()} affect the return value. */
